% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\docType{data}
\name{cpt}
\alias{cpt}
\title{OSS Data for CPT Codes}
\format{
three columns: \verb{cpt_code<integer>}, \verb{cpt_description<character>}, and \verb{oss<integer>}
}
\source{
From Shinall MC, Arya S, Youk A, et al. Association of Preoperative Patient Frailty and Operative Stress With Postoperative Mortality. JAMA Surg. 2020;155(1):e194620. doi:10.1001/jamasurg.2019.4620. Table obtained from \verb{sarya1 AT stanford DOT edu} and dated 2019-11-14
}
\usage{
cpt
}
\description{
A data frame containing CPT codes, descriptions of each procedure, and each
procedure's associated operative stress score (OSS). This table is used for
matching CPT codes to OSS in the \code{\link[=oss]{oss()}} function.
}
\keyword{datasets}
