% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss}
\alias{oss}
\title{Operative Stress Score}
\usage{
oss(.data, .column_prefix = NA, .ncores = 1)
}
\arguments{
\item{.data}{the dataset containing units' CPT Code data for which the OSS scores
are to be calculated. N.B. This parameter is NOT the table \link{cpt} which
contains the key:value mapping of CPT codes to OSS. Example values for \code{.data}
are in \link{example_cpts}}

\item{.column_prefix}{prefix in names of columns which contain CPT codes to
be assigned OSS values (e.g. 'procd' for \code{procd1}, \code{procd2}, etc.)}

\item{.ncores}{the number of cores to use in parallel processing when generating
OSS columns for each column of CPT codes}
}
\value{
the same dataset passed to the function with new columns for OSS and
procedure description corresponding to each column specified by
\code{.column_prefix}. If there is no OSS value matching a given CPT, the OSS and
desctription returned are both \code{NA}
}
\description{
Calculate the OSS for CPT Codes contained in a table. OSS for each individual CPT
code are stored in the \link{cpt} data frame.
}
\details{
\code{.data} can be a base R \code{data.frame}, a \code{data.table}, or a tidyverse \code{tibble}. The
table should contain CPT codes grouped by either patient or encounter (henceforth
referred to as the 'unit').

\code{oss} works both for tables structured in either of two ways. The first structure is
that in which there are multiple columns for CPT codes and a single row for
each unit. This is the same data format assumed for the input data in the
\code{icdpicr} package and is the default structure in data sets distributed from
sources such as HCUP. The second structure is that in which each
CPT code is stored in a single column, and multiple rows may contain CPT code
data pertaining to a single unit. This is the same data format assumed in the
\code{comorbidity} package. Data in which multiple codes are stored in a single
cell of a data.frame (perhaps separated by spaces or delimiters, in a string)
need to be pre-processed into the either of the aforementioned data formats
before using \code{oss}.
}
\examples{
oss(example_cpts$procd) #calculate oss for vector of CPT codes
oss(example_cpts, 'procd') #calculate oss for column of CPT codes
oss(example_cpts, 'procd', .ncores = 2) #same as above, in parallel
}
