% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cpts.R
\docType{data}
\name{example_cpts}
\alias{example_cpts}
\title{Example CPT Codes}
\format{
one column: \verb{procd<integer>}
}
\usage{
example_cpts
}
\description{
A data frame containing example CPT codes, including invalid codes not
present in \link{cpt}. This table is used for the examples only.
}
\keyword{datasets}
