% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.R
\name{oss}
\alias{oss}
\title{Operative Stress Score}
\usage{
oss(.data, .column_prefix, .id = NA, .ncores = 1)
}
\arguments{
\item{.data}{the dataset containing units' CPT Code data for which the OSS scores
are to be calculated. N.B. This parameter is NOT the table \link{cpt} which
contains the key:value mapping of CPT codes to OSS. Example values for \code{.data}
include \link{example_cpts_long} and \link{example_cpts_wide}}

\item{.column_prefix}{prefix in names of columns which contain CPT codes to
be assigned OSS values (e.g. 'procd' for \code{procd1}, \code{procd2}, etc.)}

\item{.id}{name of column containing patient or encounter IDs. Default
value \code{NA} assumes no ID column and that each row refers to a unique unit. If
value is a vector of column names, then all unique combinations of values from
the columns will be used as the grouping criterion (similar to \link[dplyr:group_by]{dplyr::group_by})}

\item{.ncores}{the number of cores to use in parallel processing when generating
OSS columns for each column of CPT codes}
}
\value{
the same dataset passed to the function with new columns for OSS
corresponding to each column specified by \code{.column_prefix}
}
\description{
Calculate the OSS for CPT Codes contained in a table. OSS for each individual CPT
code are stored in the \link{cpt} data frame.
}
\details{
\code{.data} can be a base R \code{data.frame}, a \code{data.table}, or a tidyverse \code{tibble}. The
table should contain CPT codes grouped by either patient or encounter (henceforth
referred to as the 'unit').

\code{oss} works both for tables structured in either of two ways. The first structure is
that in which there are multiple columns for CPT codes and a single row for
each unit. This is the same data format assumed for the input data in the
\code{icdpicr} package and is the default structure in data sets distributed from
sources such as HCUP. The second structure is that in which each
CPT code is stored in a single column, and multiple rows may contain CPT code
data pertaining to a single unit. This is the same data format assumed in the
\code{comorbidity} package.
}
\examples{
oss(example_cpts_wide, 'procd') #no need to specify .id given each row is assumed to represent a unique unit
oss(example_cpts_long, 'procd', id = 'mrn') #multiple rows with same value of `mrn` all specify the same unit
oss(example_cpts_long, 'procd', id = 'mrn', .ncores = 4) #compute OSS in parallel across 4 cores

}
